backend/routes/auth.py
from fastapi import APIRouter, HTTPException, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from fastapi.responses import JSONResponse
from pydantic import BaseModel
from jose import JWTError, jwt
from datetime import datetime, timedelta
import hashlib, secrets, os

router = APIRouter()

SECRET_KEY = os.getenv("JWT_SECRET", "smartresume-secret-key-change-in-prod")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24 * 7  # 7 days

security = HTTPBearer(auto_error=False)

# In-memory store (replace with DB in production)
users_store: dict = {}

class RegisterRequest(BaseModel):
    name: str
    email: str
    password: str

class LoginRequest(BaseModel):
    email: str
    password: str

# Simple but reliable password hashing — no bcrypt dependency issues
def hash_password(pwd: str) -> str:
    salt = secrets.token_hex(16)
    hashed = hashlib.sha256((salt + pwd).encode()).hexdigest()
    return f"{salt}:{hashed}"

def verify_password(pwd: str, stored: str) -> bool:
    try:
        salt, hashed = stored.split(":", 1)
        return hashlib.sha256((salt + pwd).encode()).hexdigest() == hashed
    except Exception:
        return False

def create_token(email: str) -> str:
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    return jwt.encode({"sub": email, "exp": expire}, SECRET_KEY, algorithm=ALGORITHM)

def verify_token(token: str) -> str:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email = payload.get("sub")
        if not email:
            raise HTTPException(status_code=401, detail="Invalid token")
        return email
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid or expired token")

def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)):
    if not credentials:
        raise HTTPException(status_code=401, detail="Not authenticated")
    email = verify_token(credentials.credentials)
    user = users_store.get(email)
    if not user:
        raise HTTPException(status_code=401, detail="User not found")
    return user

@router.post("/register")
def register(body: RegisterRequest):
    if not body.name or not body.email or not body.password:
        raise HTTPException(status_code=400, detail="All fields are required")
    if len(body.password) < 6:
        raise HTTPException(status_code=400, detail="Password must be at least 6 characters")
    if body.email in users_store:
        raise HTTPException(status_code=400, detail="Email already registered")

    users_store[body.email] = {
        "name": body.name,
        "email": body.email,
        "password_hash": hash_password(body.password),
    }
    token = create_token(body.email)
    return JSONResponse({"token": token, "name": body.name, "email": body.email})

@router.post("/login")
def login(body: LoginRequest):
    user = users_store.get(body.email)
    if not user or not verify_password(body.password, user["password_hash"]):
        raise HTTPException(status_code=401, detail="Invalid email or password")
    token = create_token(body.email)
    return JSONResponse({"token": token, "name": user["name"], "email": user["email"]})

@router.get("/me")
def get_me(current_user: dict = Depends(get_current_user)):
    return {"name": current_user["name"], "email": current_user["email"]}
