backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
python-multipart==0.0.17
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
reportlab==4.2.5
python-jose[cryptography]==3.3.0
httpx==0.27.2
